Shift blink detection code to a shared component
