+++ forge-demo/package.json
@@ -24,7 +24,7 @@
   "license": "MIT",
   "devDependencies": {
     "@babel/core": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
+    "@babel/preset-react": "^7.16.0",
     "@electron-forge/cli": "^6.0.0-beta.61",
     "@electron-forge/maker-deb": "^6.0.0-beta.61",
     "@electron-forge/maker-rpm": "^6.0.0-beta.61",
@@ -39,6 +39,7 @@
     "style-loader": "^3.0.0"
   },
   "dependencies": {
+    "@babel/eslint-parser": "^7.16.0",
     "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
     "@fontsource/roboto": "^4.5.1",
@@ -51,6 +52,8 @@
     "copy-webpack-plugin": "^9.0.1",
     "electron-squirrel-startup": "^1.0.0",
     "electron-store": "^8.0.1",
+    "eslint": "^8.1.0",
+    "eslint-plugin-react": "^7.26.1",
     "lodash": "^4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
