+++ forge-demo/src/react-app/main-screen.jsx
@@ -1,8 +1,17 @@
+import { Paper } from "@mui/material";
+import { Box } from "@mui/system";
 import React from "react";
+import { CameraWithHighlights } from "./camera-with-highlights.jsx";
 import { useResizer } from "./hooks/use-resizer";
 
 export const MainScreen = () => {
   useResizer({ width: 800, height: 600 });
 
-  return <h1>This is the main screen</h1>;
+  return (
+    <Box>
+      <Paper>
+        <CameraWithHighlights />
+      </Paper>
+    </Box>
+  );
 };
